Move ParameterType to it's own class
