+++ core/src/main/java/org/frankframework/senders/XsltSender.java
@@ -32,8 +32,8 @@ import org.frankframework.core.SenderException;
 import org.frankframework.core.SenderResult;
 import org.frankframework.jta.IThreadConnectableTransactionManager;
 import org.frankframework.parameters.Parameter;
-import org.frankframework.parameters.Parameter.ParameterType;
 import org.frankframework.parameters.ParameterList;
+import org.frankframework.parameters.ParameterType;
 import org.frankframework.parameters.ParameterValueList;
 import org.frankframework.stream.IThreadCreator;
 import org.frankframework.stream.Message;
