+++ core/src/test/java/org/frankframework/parameters/ParameterTest.java
@@ -27,7 +27,6 @@ import org.frankframework.core.PipeLine.ExitState;
 import org.frankframework.core.PipeLineExit;
 import org.frankframework.core.PipeLineResult;
 import org.frankframework.core.PipeLineSession;
-import org.frankframework.parameters.Parameter.ParameterType;
 import org.frankframework.pipes.PutInSession;
 import org.frankframework.processors.CorePipeLineProcessor;
 import org.frankframework.processors.CorePipeProcessor;
