+++ core/src/test/java/org/frankframework/senders/XsltSenderParameterNamespaceAwarenessTest.java
@@ -3,7 +3,7 @@ package org.frankframework.senders;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.frankframework.parameters.Parameter;
-import org.frankframework.parameters.Parameter.ParameterType;
+import org.frankframework.parameters.ParameterType;
 import org.frankframework.stream.Message;
 import org.frankframework.testutil.TestFileUtils;
 import org.junit.jupiter.api.Test;
